# docs/swagger.yaml
definitions:
  models.AuthInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.Token:
    properties:
      token:
        type: string
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      updatedAt:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https:://swagger.io/resources/open-api/
host: localhost:3000
info:
  contact:
    email: daniel.kwame.okyere101@gmail.com
    name: Daniel Okyere
    url: http://www.swagger.io/support
  description: This project is social media api
  title: Go Social Media API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates user
      parameters:
      - description: Login details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AuthInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Token'
      summary: Authenticates use and provides JWT
      tags:
      - authentication
  /register:
    post:
      consumes:
      - application/json
      description: Simple discription to function
      parameters:
      - description: Signup information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AuthInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Creates a user and persist to database
      tags:
      - authentication
  /user/me:
    get:
      consumes:
      - application/json
      description: Get the currently authenticated user
      parameters:
      - default: Bearer YOUR_TOKEN
        description: Authorization Header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Get user profile
      tags:
      - User
swagger: "2.0"